caption style refactor code
