+++ backend/app.py
@@ -111,7 +111,7 @@ class LunarisApp:
     def setup_resend(self):
         resend.api_key = os.environ.get('RESEND_API_KEY')
 
-    def process_video_thread(self, video_link, video_path, project_id, clerk_user_id, user_email, video_quality, video_type, start_time, end_time, clip_length, keywords):
+    def process_video_thread(self, video_link, video_path, project_id, clerk_user_id, user_email, video_quality, video_type, start_time, end_time, clip_length, keywords, caption_style):
         with self.app.app_context():
             try:
                 self.app.logger.info(f"Starting video processing for project: {project_id}")
@@ -135,7 +135,8 @@ class LunarisApp:
                 processed_clip_ids = self.video_processor.crop_and_add_subtitles(
                     downloaded_video_path, 
                     interesting_data, 
-                    video_type, 
+                    video_type,
+                    caption_style,
                     self.output_path, 
                     s3_client=self.s3_client, 
                     s3_bucket=self.s3_bucket, 
@@ -250,6 +251,8 @@ class LunarisApp:
 
         self.app.logger.info(f'Received video: {video_link or video_path}')
 
+        caption_style = request.form.get('captionStyle', 'elon')  # Default to 'elon' if not provided
+
         thread = Thread(target=self.process_video_thread, args=(video_link, 
                                                                 video_path,
                                                                 project_id, 
@@ -263,7 +266,8 @@ class LunarisApp:
                                                                     'min': float(request.form.get('clipLengthMin')),
                                                                     'max': float(request.form.get('clipLengthMax'))
                                                                 }, 
-                                                                request.form.get('keywords')))
+                                                                request.form.get('keywords'),
+                                                                caption_style))
         thread.start()
         
         return jsonify({'message': 'Video processing started', 'project_id': str(project_id)}), 202
